3, 4

+++ src/main/java/ru/itmo/webmail/model/repository/UserRepository.java
@@ -9,5 +9,5 @@ public interface UserRepository {
     User findByLogin(String login);
     User findByEmail(String email);
     List<User> findAll();
-    int findCount();
+    long findCount();
 }
